Introduce the `base` library
Originally, `common` library  contained the routines used by multiple
harnesses. These days LKL was always considered an option. That became
an issue when custom mutator was implemented as the `common` library
turned out not really common due to its nonstandard dependency.

This commit introduces the `base` library that is intended to be usable
by harnesses, the custom mutator and anything else introduced in the
future.

The `common` library has to be renamed later, but let's choose a
different name for the new library anyway, to be sure.